requirements.txt
asgiref==3.4.1
Django==4.0.1
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-colorfield==0.6.3
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.13.4
django-widget-tweaks==1.4.12
gunicorn==20.1.0
Pillow==9.0.0
six==1.16.0
sqlparse==0.4.2
